requirements.txt
flask==2.3.2
twilio==8.12.0
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
pyjiit